Moved locator settings to core
